cli: remove `config dump` and add `fde` subcommand
In this commit, we added the `fde` subcommand with those two functions:
- `show-reference-value` to generate the reference value required by rvps
- `dump-config` to show the fde config bundle in toml format

Both of those functions can be used on current running system or on a standalone disk file / disk device (with the --disk option).

The fde config bundle is a unified toml data struct which is basically the content of both global.toml and fde.toml in ${config_dir}. It defines the behaviors of cryptpilot in initrd.

This data struct is measured and you can dump it from a confidential disk. The output of `dump-config` is started with a special prefix `#cryptpilot-fde-config` (in previous version we have another prefix `#cryptpilot-config` which contains all the config including the volumes config, but it is now deprecated), and can be deployed via cloudinit userdata field.

Signed-off-by: Kun Lai <laikun@linux.alibaba.com>