cryptpilot-convert: compatible with different types of disk partitions
Disks are divided into different combinations such as whether there is efi partition or not,
and whether there is boot partition or not. Previously, the system lvm partition creation
range was extended all the way to the end of the disk device. If there were partitions after
the rootfs partition, they would be damaged. It has been repaired